move schema into submission-form folder and payload mapping into onSubmit
