+++ bikespace_frontend/src/components/submission/submission-form-controller/submission-form-controller.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {render, screen} from '@testing-library/react';
 import {FormProvider, useForm} from 'react-hook-form';
 
-import {SubmissionSchema} from '../schema';
+import {SubmissionSchema} from '../submission-form/schema';
 
 import {SubmissionFormController} from './SubmissionFormController';
